Fixes deprecated syscall issue on macOS 10.12 (https://github.com/mfreiholz/humblelogging/issues/23)
* Visual Studio 2015 adds support for %zu (see change in src/formatter/patternformatter.cpp)
* I recommend storing the ThreadID as size_t or an equivalent since that appears to be the popular approach

+++ include/humblelogging/logevent.h
@@ -20,7 +20,7 @@ public:
 	const std::string& getFunction() const { return _function; }
 	const time_t& getTime() const { return _time; }
 	unsigned int getPid() const { return _pid; }
-	unsigned int getTid() const { return _tid; }
+	size_t getTid() const { return _tid; }
 
 private:
 	int _level;
@@ -30,7 +30,7 @@ private:
 	std::string _function;
 	time_t _time;
 	unsigned int _pid;
-	unsigned int _tid;
+	size_t _tid;
 };
 
 HL_NAMESPACE_END
