+++ include/humblelogging/util/processinfo.h
@@ -2,6 +2,7 @@
 #define HL_PROCESSINFO_H
 
 #include "humblelogging/defines.h"
+#include <cstddef>
 
 HL_NAMESPACE_BEGIN
 
@@ -9,7 +10,7 @@ class HUMBLE_EXPORT_API ProcessInfo
 {
 public:
 	static unsigned int getProcessID();
-	static unsigned int getThreadID();
+	static size_t getThreadID();
 };
 
 HL_NAMESPACE_END
